aider: Define `install_pkgs` role as a dependency in `gnome_settings` role's `meta/main.yml` and pass `gnome_packages` variable to it.
# Aider chat conversation:

USER: Does it have to be in the tasks main.yml? Can we put it in the meta main.yml?

ASSISTANT: Yes, you can define role dependencies in the `meta/main.yml` file of an Ansible role. This is actually the recommended place to define role dependencies because it allows Ansible to automatically resolve and include the dependent roles when the parent role is executed.

If you want to include the `install_pkgs` role as a dependency of another role and pass the `packages` variable to it, you would define this in the `meta/main.yml` file of the dependent role. Here's how you can do it:

```diff
--- vm-setup/roles/gnome_settings/meta/main.yml
+++ vm-setup/roles/gnome_settings/meta/main.yml
@@ ... @@
 dependencies:
 - role: install_pkgs
   vars:
-    packages: "{{ packages }}"
+    packages: "{{ gnome_packages }}"
```

In this example, the `gnome_settings` role would depend on the `install_pkgs` role, and the `gnome_packages` variable (which should be defined in the `gnome_settings` role's `defaults/main.yml` or `vars/main.yml`) would be passed to the `install_pkgs` role as the `packages` variable.

This way, when the `gnome_settings` role is included in a playbook, Ansible will automatically include the `install_pkgs` role as well and use the `gnome_packages` list to install the required packages.