aider: Set the `tilix_profile_filename` variable to the fqfn using the `playbook_dir` variable.
# Aider chat conversation:

USER: How can I pass the fqfn of a file in the same directory as test-role.yml to the gnome-settings role, which is expecting a fqfn?

ASSISTANT: To pass the fully qualified file name (fqfn) of a file in the same directory as `test-role.yml` to the `gnome-settings` role, you can use the `playbook_dir` variable provided by Ansible. This variable contains the path to the directory of the currently running playbook. You can then append the file name to this path to create the fqfn.

Here's the diff for the `test-role.yml` file to include the fqfn for the `tilix_profile_filename` variable:

```diff
--- vm-setup/test-role.yml
+++ vm-setup/test-role.yml
@@ ... @@
   roles:
     - role: gnome_settings
-      vars:
-        tilix_profile_filename: "harleypig-tilix-profile.dconf"
+      tilix_profile_filename: "{{ playbook_dir }}/harleypig-tilix-profile.dconf"
```

This change sets the `tilix_profile_filename` variable to the fqfn by using the `playbook_dir` variable and appending the file name `harleypig-tilix-profile.dconf` to it.