aider: Changed the background color to be 10% darker blue.
# Aider chat conversation:

USER: So now, the install packages role has a become in the tasks/main.yml ... do I still need to use the -K switch when calling the playbook?

ASSISTANT: The `-K` switch (or `--ask-become-pass`) in Ansible is used to prompt the user for the privilege escalation password (usually the sudo password) when the playbook or any roles/tasks within it use `become` to escalate privileges. If the `install_pkgs` role contains tasks that use `become`, and you are not running the playbook as a user with passwordless sudo configured, you will still need to use the `-K` switch to provide the sudo password.

If you have tasks that require privilege escalation to run as root, and you do not have passwordless sudo set up for the user running the playbook, you should use the `-K` switch to ensure that Ansible can escalate privileges when necessary.

Here is an example of how you might call the playbook with the `-K` switch:

```bash
ansible-playbook playbook.yml -K
```

This will prompt you for the sudo password at the start of the playbook run. If you have configured passwordless sudo for the user, you do not need to use the `-K` switch.

USER: Change the background color to be 10% darker blue.

ASSISTANT: To make the background color 10% darker, we need to adjust the color value in the `desktop_background` variable within the `gnome_settings` role's `defaults/main.yml` file. The current color is `#00008B` which is a medium dark blue. To make it 10% darker, we can use a color manipulation tool or a programming library to calculate the new color.

However, since I cannot execute code, I will provide a manual estimation. Darkening a color by 10% typically involves reducing the brightness of the color. For hexadecimal colors, this can be done by reducing the value of each color component (red, green, and blue) by 10%. For the color `#00008B` (which has RGB components of 0, 0, 139), we can reduce the blue component by about 10% (which is approximately 14, given that 10% of 139 is 13.9). The new blue component would be 139 - 14 = 125, which in hexadecimal is `7D`.

Here is the diff for the `gnome_settings` role's `defaults/main.yml` file:

```diff
--- vm-setup/roles/gnome_settings/defaults/main.yml
+++ vm-setup/roles/gnome_settings/defaults/main.yml
@@ ... @@
 desktop_background:
   - picture-options 'none'
   - picture-uri ''
   - picture-uri-dark ''
-  - primary-color '#00008B'
+  - primary-color '#00007D'
   - show-desktop-icons false
```

This change sets the primary color to a 10% darker shade of the original blue. Please note that color perception is complex, and this manual calculation may not perfectly represent a 10% darker color in all contexts. For precise color adjustments, it's recommended to use a color manipulation tool or library.