aider: Added optional `Path` entry to autostart `.desktop` file for specifying application starting location.
# Aider chat conversation:

USER: Is there a way to make an auto started application start in a specified location?

ASSISTANT: To make an auto-started application start in a specified location, you can modify the `.desktop` file used for autostart to include a `Path` entry, which specifies the working directory for the application.

Let's modify the `vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2` template to include an optional `Path` entry. If the `autostart_app_path` variable is set, it will be included in the `.desktop` file; otherwise, it will be omitted.

Here's the diff for the changes:

```diff
--- vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
+++ vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
@@ ... @@
 [Desktop Entry]
 Type=Application
 Exec={{ autostart_app_exec }}
+Path={{ autostart_app_path | default('') }}
 Hidden={{ autostart_app_hidden }}
 X-GNOME-Autostart-enabled=true
 Name={{ autostart_app_name }}
 Comment={{ autostart_app_comment }}
 Terminal={{ autostart_app_terminal }}
```

With this change, you can now specify the `autostart_app_path` variable when including the `create_autostart.yml` task to set the desired starting location for the application. If the variable is not provided, the `Path` entry will be left empty, which is the default behavior that does not specify a working directory.