aider: Add `wmctrl` package to the `gnome_packages` list in `gnome_settings/vars/main.yml`.
# Aider chat conversation:

USER: Add that package to the gnome settings vars/main.yml.

ASSISTANT: Here's the diff to add `wmctrl` to the `gnome_packages` list in the `vm-setup/roles/gnome_settings/vars/main.yml` file:

```diff
--- vm-setup/roles/gnome_settings/vars/main.yml
+++ vm-setup/roles/gnome_settings/vars/main.yml
@@ ... @@
 gnome_packages:
   - gnome-shell-extension-desktop-icons-ng
   - gnome-shell-extensions
   - gnome-tweaks
   - tilix
+  - wmctrl
```

This change will ensure that `wmctrl` is included in the list of packages to be installed as part of the GNOME settings role.