aider: Calculate the position to move the Tilix window to the bottom right corner of the screen dynamically.
# Aider chat conversation:

USER: Isn't there a way to find the screen resolution and the current settings of the window and calculate the x,y values?

ASSISTANT: Yes, you can indeed calculate the position to move the Tilix window to the bottom right corner of the screen dynamically by obtaining the screen resolution and the window size. To do this, you can use `xrandr` to get the screen resolution and `wmctrl` to get the window size and position.

I will modify the `move_tilix_to_bottom_right.sh` script to calculate the position dynamically:

```diff
--- vm-setup/roles/gnome_settings/files/move_tilix_to_bottom_right.sh
+++ vm-setup/roles/gnome_settings/files/move_tilix_to_bottom_right.sh
@@ -3,7 +3,20 @@
 # Move the Tilix window to the bottom right corner of the screen.
 # Replace the 0,1000,500 with the actual screen coordinates for the bottom right corner.

-TILIX_WINDOW=$(wmctrl -l | grep Tilix | awk '{print $1}')
+SCREEN_RESOLUTION=$(xrandr --current | grep '*' | uniq | awk '{print $1}')
+SCREEN_WIDTH=$(echo $SCREEN_RESOLUTION | cut -d 'x' -f 1)
+SCREEN_HEIGHT=$(echo $SCREEN_RESOLUTION | cut -d 'x' -f 2)
+
+TILIX_WINDOW_ID=$(wmctrl -l | grep Tilix | awk '{print $1}')
+TILIX_WINDOW_GEOMETRY=$(wmctrl -lG | grep "$TILIX_WINDOW_ID" | awk '{print $3 " " $4 " " $5 " " $6}')
+TILIX_WINDOW_WIDTH=$(echo $TILIX_WINDOW_GEOMETRY | cut -d ' ' -f 3)
+TILIX_WINDOW_HEIGHT=$(echo $TILIX_WINDOW_GEOMETRY | cut -d ' ' -f 4)
+
+TARGET_X=$((SCREEN_WIDTH - TILIX_WINDOW_WIDTH))
+TARGET_Y=$((SCREEN_HEIGHT - TILIX_WINDOW_HEIGHT))
+
+if [ -n "$TILIX_WINDOW_ID" ]; then
+    wmctrl -i -r "$TILIX_WINDOW_ID" -e 0,$TARGET_X,$TARGET_Y,-1,-1
+fi
```

This script now dynamically calculates the position to move the Tilix window to the bottom right corner of the screen based on the current screen resolution and the size of the Tilix window. It uses `xrandr` to get the screen resolution and `wmctrl` to get the window size and then calculates the target X and Y coordinates.