aider: Update references to the script name in `create_autostart.yml` file to `move_tilix` and ensure the script file is renamed accordingly.
# Aider chat conversation:

USER: change the name of the script to move_tilix (no extension). I will move the file in the files directory myself.

ASSISTANT: Here is the diff to update the references to the script name in the `create_autostart.yml` file:

```diff
--- vm-setup/roles/gnome_settings/tasks/create_autostart.yml
+++ vm-setup/roles/gnome_settings/tasks/create_autostart.yml
@@ ... @@
-    autostart_app_exec: "{{ ansible_env.HOME }}/bin/move_tilix_to_bottom_right.sh"
+    autostart_app_exec: "{{ ansible_env.HOME }}/bin/move_tilix"
@@ ... @@
-    path: "{{ ansible_env.HOME }}/bin/move_tilix_to_bottom_right.sh"
+    path: "{{ ansible_env.HOME }}/bin/move_tilix"
```

Make sure to rename the actual script file in the filesystem to `move_tilix` as you mentioned you would handle that part yourself.