aider: Add task to copy `move_tilix` script to `~/.local/bin` in `gnome_settings` role's `tilix.yml`.
# Aider chat conversation:

USER: How is the file moved from the files dir to the install location?

ASSISTANT: The file `move_tilix` is moved from the `files` directory to the installation location `~/.local/bin` using an Ansible task. This task would typically involve the `copy` module in Ansible to transfer the file from the local directory to the target location on the remote system.

Here is an example of how the Ansible task might look in a playbook or a role:

```yaml
- name: Copy move_tilix script to the local bin directory
  copy:
    src: move_tilix
    dest: "{{ ansible_env.HOME }}/.local/bin/move_tilix"
    mode: '0755'
```

This task would be part of an Ansible role or playbook that is executed against the target machine. The `copy` module takes the source file `move_tilix` from the control machine (where Ansible is run) and copies it to the destination path on the target machine, setting the file mode to `0755` to make it executable.

If you would like to see the changes to the Ansible playbook or role that would implement this task, please provide the specific file you would like to modify or add it to the chat if it's not already listed.

USER: Add this to the gnome settings tilix task.

ASSISTANT: Here is the diff to add the task for copying the `move_tilix` script to the `~/.local/bin` directory in the `gnome_settings` role's `tilix.yml` task file:

```diff
--- vm-setup/roles/gnome_settings/tasks/tilix.yml
+++ vm-setup/roles/gnome_settings/tasks/tilix.yml
@@ ... @@

 - include_tasks: tilix_profile.yml

+- name: Copy move_tilix script to the local bin directory
+  copy:
+    src: move_tilix
+    dest: "{{ ansible_env.HOME }}/.local/bin/move_tilix"
+    mode: '0755'
+
 - name: Autostart Tilix on login
   include_tasks: create_autostart.yml
   vars:
     autostart_app_name: "Tilix"
     autostart_app_exec: "tilix"
     autostart_app_comment: "Tilix terminal emulator"
     autostart_app_terminal: false
     autostart_app_hidden: false
```

This change adds a new task to the `tilix.yml` file that will copy the `move_tilix` script from the role's `files` directory to the `~/.local/bin` directory on the target system and make it executable.